[Improve][Zeta] Separate checkpoint benchmark pipeline
